# services/web-app/openapi.yaml
openapi: 3.0.3
info:
  title: RespiraAlly API
  description: API for the RespiraAlly respiratory care management system
  version: 1.0.0
  contact:
    name: RespiraAlly Team
    email: support@respiraally.com
servers:
  - url: http://localhost:5000/api/v1
    description: Development server
  - url: https://api.respiraally.com/api/v1
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: "INVALID_INPUT"
            message:
              type: string
              example: "Request body cannot be empty"
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        account:
          type: string
          example: "admin"
        first_name:
          type: string
          example: "思敏"
        last_name:
          type: string
          example: "王"
        email:
          type: string
          example: "sumin.wang@example.com"
        is_staff:
          type: boolean
          example: true
        is_admin:
          type: boolean
          example: false
...
    HealthProfile:
      type: object
      properties:
        height_cm:
          type: integer
          example: 160
        weight_kg:
          type: integer
          example: 55
        smoke_status:
          type: string
          enum: [never, former, current]
          example: "never"
        updated_at:
          type: string
          format: date-time
          example: "2025-01-15T10:30:00Z"
...
paths:
...
  /auth/login:
    post:
      summary: "呼吸治療師登入"
      description: "供呼吸治療師使用帳號密碼進行登入"
      tags: ["Authentication"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [account, password]
              properties:
                account:
                  type: string
                  example: "admin"
                password:
                  type: string
                  format: password
                  example: "admin"
      responses:
        '200':
          description: "登入成功"
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      token:
                        type: string
                        example: "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
                      expires_in:
                        type: number
                        example: 3600
                      user:
                        $ref: '#/components/schemas/User'
        '400':
          description: "請求格式錯誤"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: "帳號或密碼錯誤"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/line/login:
    post:
      summary: "患者 LIFF 登入"
      description: "供已註冊的患者在 LIFF 環境中，使用 lineUserId 進行登入"
      tags: ["Authentication"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [lineUserId]
              properties:
                lineUserId:
                  type: string
                  example: "U123456789abcdef123456789abcdef"
      responses:
        '200':
          description: "登入成功"
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      token:
                        type: string
                      expires_in:
                        type: number
                      user:
                        type: object
                        properties:
                          id:
                            type: integer
                          line_user_id:
                            type: string
                          first_name:
                            type: string
                          last_name:
                            type: string
                          health_profile:
                            $ref: '#/components/schemas/HealthProfile'
        '400':
          description: "lineUserId 缺失"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: "使用者未註冊"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/line/register:
    post:
      summary: "患者 LIFF 註冊"
      description: "供新患者在 LIFF 環境中，使用 lineUserId 並填寫基本資料以完成註冊"
      tags: ["Authentication"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [lineUserId, first_name, last_name]
              properties:
                lineUserId:
                  type: string
                  example: "U_new_user_abcdef123456"
                first_name:
                  type: string
                  example: "美麗"
                last_name:
                  type: string
                  example: "陳"
                gender:
                  type: string
                  example: "female"
                phone:
                  type: string
                  example: "0987654321"
                height_cm:
                  type: integer
                  example: 160
                weight_kg:
                  type: integer
                  example: 55
                smoke_status:
                  type: string
                  example: "never"
      responses:
        '201':
          description: "註冊成功並自動登入"
        '400':
          description: "缺少必要欄位"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: "使用者已存在"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Patient Management endpoints
...